Create class `Collect` and `CollectPool`
