+++ test/manual.ts
@@ -7,6 +7,7 @@ import { Server } from "../src/server";
 import { CollectorLoader } from '../src/collectors/collectorLoader';
 import { LoggableError } from '../src/error';
 import { Secret } from '../src/secret_manager/abstractSecretManager';
+import { Collect } from '../src/collect/collect';
 
 (async () => {
     let id;
@@ -60,7 +61,8 @@ import { Secret } from '../src/secret_manager/abstractSecretManager';
         }
 
         // Collect invoices
-        const { invoices, cookies } = await collector.collect_new_invoices(secret, true, [], Server.DEFAULT_LOCALE, {country: "FR", lat: '', lon: ''});
+        const collect = new Collect("")
+        const { invoices, cookies } = await collect.collect_new_invoices(collector, secret, true, [], Server.DEFAULT_LOCALE, {country: "FR", lat: '', lon: ''});
         console.log(`${invoices.length} invoices downloaded`);
 
         for (const invoice of invoices) {
